% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_helpers.R
\name{authors_footnote}
\alias{authors_footnote}
\title{Create author footnote for gt theme
@keywords internal
@param msd_source source metadata recovered from \code{\link[=msd_period]{msd_period()}}
@returns a string
@family gt helpers}
\usage{
authors_footnote(msd_source)
}
\description{
Create author footnote for gt theme
@keywords internal
@param msd_source source metadata recovered from \code{\link[=msd_period]{msd_period()}}
@returns a string
@family gt helpers
}
