% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_helpers.R
\name{bold_rowgroup}
\alias{bold_rowgroup}
\title{Make all text larger
Bold Agency names - used to increase stroke on row group label}
\usage{
bold_rowgroup(gt_obj, wt = 700)
}
\arguments{
\item{gt_obj}{gt object pass through}

\item{wt}{size (0-1000) of embiggening}
}
\value{
a modified gt object
}
\description{
Make all text larger
Bold Agency names - used to increase stroke on row group label
}
\examples{
\dontrun{
 mtcars \%>\% gt(groupname_col = "cyl") \%>\% bold_rowgroup(wt = 500)
 }
}
\seealso{
Other gt helpers: 
\code{\link{authors_footnote}()},
\code{\link{caveats_footnote}()},
\code{\link{change_footnote}()},
\code{\link{create_pd}()},
\code{\link{dedup_footnote}()},
\code{\link{delta_footnote}()},
\code{\link{extract_num_colnames}()},
\code{\link{legend}},
\code{\link{legend_q1}},
\code{\link{legend_q2}},
\code{\link{legend_q3}},
\code{\link{legend_snapshot}},
\code{\link{past_fy}()},
\code{\link{present_fy}()},
\code{\link{present_qtr}()},
\code{\link{vlc_footnote}()}
}
\concept{gt helpers}
