% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_growth.R
\name{calc_growth}
\alias{calc_growth}
\title{Calculate growth between two metrics that are positive}
\usage{
calc_growth(x, y)
}
\arguments{
\item{x}{numerator to calculate growth - most recent metric}

\item{y}{denominator to calculate growth - lagged metric}
}
\description{
Helper function for calculating growth.
Returns NA_real_ if the
}
\examples{
\dontrun{
 calc_growth(110, 100) #should give 10\%
 df \%>\% mutate(z_change = calc_growth(results_cumulative, dplyr::lag(results_cumulative, n = 4)))}

}
\seealso{
Other MDB helper functions: 
\code{\link{format_indicator}()},
\code{\link{label_aggregation}()},
\code{\link{make_achv_shape}()},
\code{\link{make_chg_shape}()}
}
\concept{MDB helper functions}
