% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_helpers.R
\name{caveats_footnote}
\alias{caveats_footnote}
\title{Create caveat note for gt theme
@keywords internal
@family gt helpers}
\usage{
caveats_footnote()
}
\description{
Create caveat note for gt theme
@keywords internal
@family gt helpers
}
