% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_helpers.R
\name{change_footnote}
\alias{change_footnote}
\title{Create change note for gt theme
@keywords internal
@family gt helpers}
\usage{
change_footnote()
}
\description{
Create change note for gt theme
@keywords internal
@family gt helpers
}
