% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse_base_tbl.R
\name{collapse_base_tbl}
\alias{collapse_base_tbl}
\title{Collapse base msd down to a desired level}
\usage{
collapse_base_tbl(df, indic_list, ...)
}
\arguments{
\item{df}{MSD or Genie data frame}

\item{indic_list}{list of indicators to be filter on}

\item{...}{dot-dot-dot that can be passed to the group_by step}
}
\description{
Helper function to collapse data down by a provided grouping
Used in the creation of the main MDB table.
Called in the creation of the treatment
}
