% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse_vlc_tbl.R
\name{collapse_vlc_tbl}
\alias{collapse_vlc_tbl}
\title{Collapse base msd down to a desired level for core treatment indicators}
\usage{
collapse_vlc_tbl(df, ...)
}
\arguments{
\item{df}{MSD or Genie data frame}

\item{...}{dot-dot-dot to be used in the grouping option}
}
\value{
collapsed data frame of TX_CURR and TX_PVLS indicators
}
\description{
Helper function to collapse data down by a provided grouping
Used in the creation of the treatment / viral load coverage MDB table
Similar to the \code{\link[=collapse_base_tbl]{collapse_base_tbl()}} function but specifically for treatment MDB table.
}
