% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_helpers.R
\name{create_pd}
\alias{create_pd}
\title{Return msd metadata for pd object}
\usage{
create_pd(df)
}
\arguments{
\item{df}{MSD or Genie extract}
}
\value{
object of the format FY@Q@
}
\description{
The derived pd object is used throughout the generation of the tables.
It is used multiple times for formatting columns and column spanners.
}
\seealso{
Other gt helpers: 
\code{\link{extract_num_colnames}()}
}
\concept{gt helpers}
