% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_helpers.R
\name{delta_footnote}
\alias{delta_footnote}
\title{Create delta note for gt theme
@keywords internal
@family gt helpers}
\usage{
delta_footnote()
}
\description{
Create delta note for gt theme
@keywords internal
@family gt helpers
}
