% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_helpers.R
\name{embiggen}
\alias{embiggen}
\title{Embiggen parts of mdb table
A noble spirit embiggens the smallest man}
\usage{
embiggen(gt_obj, tbl_size = 15, ftnote_size = 10, source_size = 10)
}
\arguments{
\item{gt_obj}{gt object to be embiggened}

\item{tbl_size}{font size for the core table}

\item{ftnote_size}{font size for the footnotes}

\item{source_size}{font size for the source notes}
}
\value{
a modified gt object
}
\description{
Embiggen parts of mdb table
A noble spirit embiggens the smallest man
}
\examples{
\dontrun{
# embiggen
mtcars \%>\% gt(groupname_col = "cyl") \%>\% embiggen(tbl_size = 15)

# de-embiggen
mtcars \%>\% gt(groupname_col = "cyl") \%>\% embiggen(tbl_size = 8)
}
}
