% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_helpers.R
\name{extract_num_colnames}
\alias{extract_num_colnames}
\title{Extract a vector of numeric column names}
\usage{
extract_num_colnames(df)
}
\arguments{
\item{df}{data frame from the \code{\link[=reshape_mdb_tx_df]{reshape_mdb_tx_df()}} call}
}
\value{
vector of column names for all numeric vars
}
\description{
This helper function is used to extract the names of all the numeric columns in the TX mdb table.
The result is passed to the treatment theme for use in formatting columns.
}
\seealso{
Other gt helpers: 
\code{\link{create_pd}()}
}
\concept{gt helpers}
