% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_indicators.R
\name{fetch_indicators}
\alias{fetch_indicators}
\title{Fetches indicators used to make tables, accounting for quarter of MSD and semi-annual indicators}
\usage{
fetch_indicators(df, tab = "main")
}
\arguments{
\item{df}{data frame on which indicators are to be fetched}

\item{tab}{which tab to fetch (main or treatment) from googlesheet}
}
\value{
dataframe of indicators
}
\description{
MDB indicators are stored on google drive. This helper function fetches a desired tab.
Used to create filters, sort factor levels, and apply plain labels to indicators.
}
\examples{
\dontrun{
 fetch_indicators(ou_im, tab = "main")
 fetch_indicators(ou_im, tab = "treatment")}
  
  
}
