% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fetch_metadata}
\alias{fetch_metadata}
\title{Fetch metadata}
\usage{
fetch_metadata()
}
\value{
A list containing metadata information.
}
\description{
Retrieves the metadata set for the package.
}
\examples{
\dontrun{
 meta <- get_metadata()
}
}
