% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99_utilities.R
\name{fmt_tbl_cols}
\alias{fmt_tbl_cols}
\title{Remove extra stubs on column names for potential use in mdb table names}
\usage{
fmt_tbl_cols(df)
}
\arguments{
\item{df}{formatted dataframe of indicator}
}
\description{
Remove extra stubs on column names for potential use in mdb table names
}
