% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_indicator.R
\name{format_indicator}
\alias{format_indicator}
\title{Formats PEPFAR indicators for MDB tables}
\usage{
format_indicator(x)
}
\arguments{
\item{x}{PEPFAR indicator name as an abbreviation and the full definition}
}
\value{
html formatted string that can be rendered in gt
}
\description{
Formats a string containing the main PEPFAR abbrevation and plain text.
The PEPFAR abbreviation is listed first and the plain text in a line below.
Used to create the upper portion of the MDB tables.
}
\examples{
\dontrun{
 format_indicator("HTS_TST Received HIV testing service and results") \%>\% 
 gt::html() \%>\%  htmltools::html_print()
 df \%>\% mutate(indicator2 = format_indicator(indicator2)) }
 
 
}
\seealso{
Other MDB helper functions: 
\code{\link{calc_growth}()},
\code{\link{label_aggregation}()},
\code{\link{make_achv_shape}()},
\code{\link{make_chg_shape}()}
}
\concept{MDB helper functions}
