% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_helpers.R
\docType{data}
\name{legend_q1}
\alias{legend_q1}
\title{Object pointing to github location of legend for Q1}
\format{
An object of class \code{character} of length 1.
}
\usage{
legend_q1
}
\description{
This helper object returns the location of the q1_cumulative legend.
The object is passed to a legend_chunk f() that creates md for the legend.
This can then be inserted into the subtitle as an image.
}
\seealso{
Other gt helpers: 
\code{\link{authors_footnote}()},
\code{\link{caveats_footnote}()},
\code{\link{change_footnote}()},
\code{\link{create_pd}()},
\code{\link{dedup_footnote}()},
\code{\link{delta_footnote}()},
\code{\link{extract_num_colnames}()},
\code{\link{legend_q2}},
\code{\link{legend_q3}},
\code{\link{legend_snapshot}},
\code{\link{past_fy}()},
\code{\link{present_fy}()},
\code{\link{present_qtr}()},
\code{\link{vlc_footnote}()}
}
\concept{gt helpers}
\keyword{datasets}
