% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_helpers.R
\docType{data}
\name{legend_snapshot}
\alias{legend_snapshot}
\title{Object pointing to github location of legend for snapshot indicators and Q4}
\format{
An object of class \code{character} of length 1.
}
\usage{
legend_snapshot
}
\description{
Use legend instead as quarterly legends are no longer necessary
}
\seealso{
Other gt helpers: 
\code{\link{authors_footnote}()},
\code{\link{bold_rowgroup}()},
\code{\link{caveats_footnote}()},
\code{\link{change_footnote}()},
\code{\link{dedup_footnote}()},
\code{\link{delta_footnote}()},
\code{\link{extract_num_colnames}()},
\code{\link{legend}},
\code{\link{legend_q1}},
\code{\link{legend_q2}},
\code{\link{legend_q3}},
\code{\link{past_fy}()},
\code{\link{present_fy}()},
\code{\link{present_qtr}()},
\code{\link{vlc_footnote}()}
}
\concept{gt helpers}
\keyword{datasets}
