% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_achv_shape.R
\name{make_achv_shape}
\alias{make_achv_shape}
\title{Generate colored circle as an svg}
\usage{
make_achv_shape(x)
}
\arguments{
\item{x}{color hex code created by adorn_achievement function}
}
\value{
return gt html() code for an svg
}
\description{
Description
Function to create and color achievement circles.
The circles are placed next to the most recent achievement level in the main MDB table.
}
\details{
Returns a html formatting circle that can be embedded in gt object
}
\examples{
\dontrun{
 achv_circle(glitr::scooter) \%>\% htmltools::html_print()
 df \%>\% mutate(achv_color = achv_circle(achv_color)) }
}
