% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_chg_shape.R
\name{make_chg_shape}
\alias{make_chg_shape}
\title{Create fontawesome performance arrow}
\usage{
make_chg_shape(change_dir)
}
\arguments{
\item{change_dir}{variable indicating direction of change or not applicable}
}
\value{
gt html() code for an svg
}
\description{
Creates an svg object based on the direction of change of a performance variable.
The svg code can be embedded in a data frame and passed to gt() for prettifying a table.
Appears to work well when applied with purrr::map()
}
\examples{
\dontrun{
 rank_chg("increase") \%>\% htmltools::html_print()
 rank_chg("decrease") \%>\% htmltools::html_print()
 df \%>\% mutate(chg_dir = purrr::map(present_z_direction, rank_chg))
 }


}
\seealso{
Other MDB helper functions: 
\code{\link{calc_growth}()},
\code{\link{format_indicator}()},
\code{\link{label_aggregation}()},
\code{\link{make_achv_shape}()}
}
\concept{MDB helper functions}
