% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_footnotes.R
\name{make_footnotes}
\alias{make_footnotes}
\title{Create recommended table footnotes}
\usage{
make_footnotes(msd_source)
}
\arguments{
\item{msd_source}{key metadata from the MSD used in the table creation}
}
\description{
Wrapper function to create footnotes that are called throughout the table creation process.
}
\examples{
\dontrun{
make_footnotes()
# paste clipboard to session and run }

}
