% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_mdb_tx_df.R
\name{make_mdb_tx_df}
\alias{make_mdb_tx_df}
\title{Creates the base table for the treatment / VLS MDB tables}
\usage{
make_mdb_tx_df(df, resolve_issues = "TRUE")
}
\arguments{
\item{df}{usually and ou_im data frame}

\item{resolve_issues}{logical indicating whether or not known issues are removed

TODO: Should this be broken out into 3 separate functions?}
}
\description{
Use the \code{\link[=collapse_vlc_tbl]{collapse_vlc_tbl()}} to combine different data frames needed to calculate VLS and VLC.
Filters for mechs with known issues by default, can be controlled with resolve_issues argument
Wrapper function to create three core tables needed for the treatment MDB data frame.

First, it creates a TX_CURR table and excludes South Africa from the USAID Global table.
Second, \code{\link[=collapse_vlc_tbl]{collapse_vlc_tbl()}} is called to create the VLS/VLC table needed for derived indicators.
Third, Three month + MMD is calculated from the TX_CURR table.
Finally, the tables are combined to be used in creation of the VLS/VLC and MMD+ Share calculations.
}
