% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_time_metadata.R
\name{make_time_metadata}
\alias{make_time_metadata}
\title{Create time objects for gt tables}
\usage{
make_time_metadata(df = ou_im)
}
\description{
Wrapper function to create numerous time objects that are called throughout the table creation process.
}
\examples{
\dontrun{
make_time_metadata()
# paste clipboard to session and run }

}
