% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdb_treatment_theme.R
\name{mdb_treatment_theme}
\alias{mdb_treatment_theme}
\title{GT theme for main MDB tables}
\usage{
mdb_treatment_theme(df, numeric_cols, ...)
}
\arguments{
\item{df}{\code{\link[=reshape_mdb_tx_df]{reshape_mdb_tx_df()}} output}

\item{numeric_cols}{an object containing the names of the columns that are numeric}

\item{...}{dot-dot-dot option to pass additional formatting to gt object}
}
\value{
formatted gt object
}
\description{
A gt theme to be applied to a reshaped data frame for the creation of MDB tables.
The theme formats all columns and rows for the main MDB tables.
A different theme exists for the main tables as the layout is different.
}
\examples{
\dontrun{
 mdb_df <- make_mdb_tx_df(ou_im, resolve_issues = F)
 mdb_tbl <- reshape_mdb_tx_df(mdb_df)
 numeric_cols <- mdb_tbl \%>\% select_if(is.numeric) \%>\% names()
 mdb_tbl \%>\% filter(operatingunit == "Zambia") \%>\% 
 gt(groupname_col = "agency") \%>\% 
 mdb_treatment_theme(numeric_cols)
 
 }
}
\seealso{
Other MDB gt themes: 
\code{\link{mdb_main_theme}()}
}
\concept{MDB gt themes}
