% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msd_period.R
\name{msd_period}
\alias{msd_period}
\title{Identify MSD Period and Type}
\usage{
msd_period(type = "OU_IM", period = NULL)
}
\arguments{
\item{type}{Type of MSD (OU_IM, PSNU, PSNU_IM, NAT_SUBNAT)}

\item{period}{can provide period from ICPIutilities::identifypd(), or it will run if NULL}
}
\value{
FY00Q0t MSD
}
\description{
requires having si_path setup from glamr
}
