% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_helpers.R
\name{past_fy}
\alias{past_fy}
\title{Create past_fy object for gt theme
@keywords internal
@param pd of the format FYXXQX that is returned from \code{\link[gophr:identifypd]{gophr::identifypd()}}
@return a string
@family gt helpers}
\usage{
past_fy(pd)
}
\description{
Create past_fy object for gt theme
@keywords internal
@param pd of the format FYXXQX that is returned from \code{\link[gophr:identifypd]{gophr::identifypd()}}
@return a string
@family gt helpers
}
