% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_helpers.R
\name{present_fy}
\alias{present_fy}
\title{Create present fy object for gt theme
@keywords internal
@param pd of the format FY@Q@ that is returned from \code{\link[gophr:identifypd]{gophr::identifypd()}}
@returns a string
@family gt helpers}
\usage{
present_fy(pd)
}
\description{
Create present fy object for gt theme
@keywords internal
@param pd of the format FY@Q@ that is returned from \code{\link[gophr:identifypd]{gophr::identifypd()}}
@returns a string
@family gt helpers
}
