% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_helpers.R
\name{present_qtr}
\alias{present_qtr}
\title{Create present quarter object for gt theme
@keywords internal
@param pd of the format FYXXQX that is returned from \code{\link[gophr:identifypd]{gophr::identifypd()}}
@returns a string
@family gt helpers}
\usage{
present_qtr(pd)
}
\description{
Create present quarter object for gt theme
@keywords internal
@param pd of the format FYXXQX that is returned from \code{\link[gophr:identifypd]{gophr::identifypd()}}
@returns a string
@family gt helpers
}
