% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshape_mdb_df.R
\name{reshape_mdb_df}
\alias{reshape_mdb_df}
\title{Prepare filtered data frame for MSD briefer formatting}
\usage{
reshape_mdb_df(df, pd = pd)
}
\arguments{
\item{df}{takes the \code{\link[=make_mdb_df]{make_mdb_df()}} results as an input}

\item{pd}{metadata from the MSD used to create time variables}
}
\value{
returns a wide formatted data frame (table) of all OUs, countries and USAID
}
\description{
\code{reshape_mdb_df} takes the output from \code{\link[=make_mdb_df]{make_mdb_df()}} and creates a gt ready data frame.
The transformed data frame retains the most recent quarter and previous year's performance.
The resulting table can be passed directly to the \code{\link[=mdb_main_theme]{mdb_main_theme()}} to create a MDB table.
Helper functions format certain columns as svgs to be rendered in the gt call.
}
\examples{
\dontrun{
 mdb_df <- make_mdb_df(ou_im, resolve_issues = F)
 mdb_tbl <- reshape_mdb_df(mdb_df)
 }
 
}
\seealso{
\code{\link[=make_mdb_df]{make_mdb_df()}} to see input required and
\code{\link[=mdb_main_theme]{mdb_main_theme()}} to see the gt theme used to format the table
\code{\link[=make_chg_shape]{make_chg_shape()}} creates an svg based on the direction of change
\code{\link[=make_achv_shape]{make_achv_shape()}} creates a colored circle based on achievement level

Other data frame munging: 
\code{\link{collapse_base_tbl}()},
\code{\link{collapse_vlc_tbl}()},
\code{\link{make_mdb_df}()},
\code{\link{make_mdb_tx_df}()},
\code{\link{reshape_mdb_tx_df}()}
}
\concept{data frame munging}
