% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_helpers.R
\name{vlc_footnote}
\alias{vlc_footnote}
\title{Create vlc change note for gt theme
@keywords internal
@family gt helpers}
\usage{
vlc_footnote()
}
\description{
Create vlc change note for gt theme
@keywords internal
@family gt helpers
}
