% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_helpers.R
\name{authors_footnote}
\alias{authors_footnote}
\title{Create author footnote for gt theme}
\usage{
authors_footnote(pd)
}
\arguments{
\item{pd}{source metadata recovered from from \code{\link[=fetch_metadata]{fetch_metadata()}}}
}
\value{
a string
}
\description{
Create author footnote for gt theme
}
\seealso{
Other gt helpers: 
\code{\link{bold_rowgroup}()},
\code{\link{caveats_footnote}()},
\code{\link{change_footnote}()},
\code{\link{dedup_footnote}()},
\code{\link{delta_footnote}()},
\code{\link{extract_num_colnames}()},
\code{\link{legend}},
\code{\link{legend_q1}},
\code{\link{legend_q2}},
\code{\link{legend_q3}},
\code{\link{legend_snapshot}},
\code{\link{past_fy}()},
\code{\link{present_fy}()},
\code{\link{present_qtr}()},
\code{\link{vlc_footnote}()}
}
\concept{gt helpers}
\keyword{internal}
