% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse_base_tbl.R
\name{collapse_base_tbl}
\alias{collapse_base_tbl}
\title{Collapse base MSD down to a desired level for main MDB table}
\usage{
collapse_base_tbl(df, indic_list, ...)
}
\arguments{
\item{df}{MSD or Genie data frame}

\item{indic_list}{list of indicators to be filter on}

\item{...}{dot-dot-dot that can be passed to the group_by step}
}
\description{
Helper function to collapse data down by a provided grouping.
Used in the creation of the main MDB table.
}
\seealso{
\link{collapse_vlc_tbl} to see equivalent collapse sequence for VLC/VLCS variables

Other data frame munging: 
\code{\link{collapse_vlc_tbl}()},
\code{\link{make_mdb_df}()},
\code{\link{make_mdb_tx_df}()},
\code{\link{reshape_mdb_df}()},
\code{\link{reshape_mdb_tx_df}()}
}
\concept{data frame munging}
