% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_helpers.R
\name{create_pd}
\alias{create_pd}
\title{Return msd metadata for pd object}
\usage{
create_pd(df)
}
\value{
a string
}
\description{
This function is depracated and has been replaced by \code{\link[=fetch_metadata]{fetch_metadata()}}

param df MSD or Genie extract
return object of the format FYXXQX
family gt helpers
}
\keyword{internal}
