% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_helpers.R
\name{dedup_footnote}
\alias{dedup_footnote}
\title{Create dedup note for gt theme
@keywords internal
@family gt helpers}
\usage{
dedup_footnote()
}
\description{
Create dedup note for gt theme
@keywords internal
@family gt helpers
}
