% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_helpers.R
\name{delta_footnote}
\alias{delta_footnote}
\title{Create delta note for gt theme}
\usage{
delta_footnote()
}
\description{
Create delta note for gt theme
}
\seealso{
Other gt helpers: 
\code{\link{authors_footnote}()},
\code{\link{bold_rowgroup}()},
\code{\link{caveats_footnote}()},
\code{\link{change_footnote}()},
\code{\link{dedup_footnote}()},
\code{\link{extract_num_colnames}()},
\code{\link{legend}},
\code{\link{legend_q1}},
\code{\link{legend_q2}},
\code{\link{legend_q3}},
\code{\link{legend_snapshot}},
\code{\link{past_fy}()},
\code{\link{present_fy}()},
\code{\link{present_qtr}()},
\code{\link{vlc_footnote}()}
}
\concept{gt helpers}
\keyword{internal}
