% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_indicators.R
\name{fetch_indicators}
\alias{fetch_indicators}
\title{Fetches indicators used to make tables, accounting for quarter of MSD and semi-annual indicators}
\usage{
fetch_indicators(pd = meta, tab = "main")
}
\arguments{
\item{pd}{period used to pull quarterly or quaterly plus semi-annual indicators}

\item{tab}{which table is being created, main or treatment}
}
\value{
dataframe of indicators
}
\description{
MDB indicators are stored on google drive. This helper function fetches a desired tab.
Used to create filters, sort factor levels, and apply plain labels to indicators.
}
\examples{
\dontrun{
 fetch_indicators(ou_im, tab = "main")
 fetch_indicators(ou_im, tab = "treatment")}
  
  
}
