% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_aggregation.R
\name{label_aggregation}
\alias{label_aggregation}
\title{Label aggregation level}
\usage{
label_aggregation(df, type = "OU")
}
\arguments{
\item{df}{MSD or genie data frame that has been ran through collapse_base_tbl()}

\item{type}{Level of aggregation to be labeled}
}
\description{
Labels the level of aggregation using regiona, agency or operating unit
Used to apply a filter column to the single data frame returned from table operations
}
\seealso{
Other MDB helper functions: 
\code{\link{calc_growth}()},
\code{\link{format_indicator}()},
\code{\link{make_achv_shape}()},
\code{\link{make_chg_shape}()}
}
\concept{MDB helper functions}
