% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_footnotes.R
\name{make_footnotes}
\alias{make_footnotes}
\title{Create recommended table footnotes}
\usage{
make_footnotes()
}
\description{
Wrapper function to create footnotes that are called throughout the table creation process.
}
\examples{
\dontrun{
make_footnotes()
# paste clipboard to session and run }

}
