% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdb_main_theme.R
\name{mdb_main_theme}
\alias{mdb_main_theme}
\title{GT theme for main MDB tables}
\usage{
mdb_main_theme(df, pd = meta, ...)
}
\arguments{
\item{df}{\code{\link[=reshape_mdb_df]{reshape_mdb_df()}} output}

\item{pd}{character object of the style FY@Q@ that is returned from \code{\link[gophr:get_metadata]{gophr::get_metadata()}}}

\item{...}{dot-dot-dot option to pass additional formatting to gt object}
}
\value{
formatted gt object
}
\description{
A gt theme to be applied to a reshaped data frame for the creation of MDB tables.
The theme formats all columns and rows for the main MDB tables.
A different theme exists for the treatment tables as the layout is different.
}
\examples{
\dontrun{
 mdb_df <- make_mdb_df(ou_im, resolve_issues = F)
 mdb_tbl <- reshape_mdb_df(mdb_df)
 mdb_tbl \%>\% filter(operatingunit == "Zambia") \%>\% gt(groupname_col = "agency") \%>\% mdb_main_theme()
 
 }
}
\seealso{
Other MDB gt themes: 
\code{\link{mdb_treatment_theme}()}
}
\concept{MDB gt themes}
