% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_helpers.R
\name{vlc_footnote}
\alias{vlc_footnote}
\title{Create vlc change note for gt theme}
\usage{
vlc_footnote()
}
\description{
Create vlc change note for gt theme
}
\seealso{
Other gt helpers: 
\code{\link{authors_footnote}()},
\code{\link{caveats_footnote}()},
\code{\link{change_footnote}()},
\code{\link{create_pd}()},
\code{\link{dedup_footnote}()},
\code{\link{delta_footnote}()},
\code{\link{extract_num_colnames}()},
\code{\link{past_fy}()},
\code{\link{present_fy}()},
\code{\link{present_qtr}()}
}
\concept{gt helpers}
\keyword{internal}
